--- components/webflow/common/constants.mjs
@@ -1,4 +1,0 @@
-export default {
-  LIMIT: 100,
-  DEPLOY_OFFSET: 25,
-};
