+++ components/webflow_v2/actions/list-orders/list-orders.mjs
@@ -1,35 +1,33 @@
-import webflow from "../../webflow.app.mjs";
+import app from "../../webflow.app.mjs";
 import constants from "../common/constants.mjs";
 
 export default {
-  key: "webflow-list-orders",
+  key: "webflow_v2-list-orders",
   name: "List Orders",
   description: "List orders. [See the docs here](https://developers.webflow.com/#get-all-orders)",
-  version: "0.0.4",
+  version: "0.0.{{ts}}",
   type: "action",
   props: {
-    webflow,
+    app,
     siteId: {
       propDefinition: [
-        webflow,
+        app,
         "sites",
       ],
     },
     status: {
       label: "Status",
-      description: "The status to filter the orders.",
+      description: "If specified, only orders with this status will be listed.",
       type: "string",
       options: constants.ORDER_STATUSES,
       optional: true,
     },
   },
   async run({ $ }) {
-    const response = await this.webflow.getOrders({
-      siteId: this.siteId,
-      status: this.status,
-    });
+    const { app, ...data } = this;
+    const response = await app.listOrders(data);
 
-    $.export("$summary", "Successfully retrieved orders");
+    $.export("$summary", `Successfully retrieved ${response?.length} orders`);
 
     return response;
   },
